refactor: colocate admin components
